# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hospital Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenInput"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed

  # ── USERS ─────────────────────────────────────────────────────────────────
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: doctorId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Appointments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentListResponse"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book a new appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Appointment booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment status/details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: cancelAppointment
      tags: [appointments]
      summary: Cancel an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Cancelled

  # ── QUEUE ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    RefreshTokenInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
...
    User:
      type: object
      required: [id, username, role, fullName, isActive, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [admin, doctor, staff]
        fullName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        registrationNumber:
          type: ["string", "null"]
        specialization:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
...
    Appointment:
      type: object
      required: [id, patientId, doctorId, appointmentDate, status, createdAt]
      properties:
        id:
          type: string
        patientId:
          type: string
        patientName:
          type: string
        patientPhone:
          type: ["string", "null"]
        doctorId:
          type: string
        doctorName:
          type: string
        appointmentDate:
          type: string
        appointmentTime:
          type: ["string", "null"]
        reason:
          type: ["string", "null"]
        status:
          type: string
          enum: [scheduled, confirmed, arrived, in_progress, completed, cancelled, no_show]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    AppointmentInput:
      type: object
      required: [patientId, doctorId, appointmentDate]
      properties:
        patientId:
          type: string
        doctorId:
          type: string
        appointmentDate:
          type: string
        appointmentTime:
          type: string
        reason:
          type: string
        notes:
          type: string
    AppointmentUpdate:
      type: object
      properties:
        appointmentDate:
          type: string
        appointmentTime:
          type: string
        status:
          type: string
          enum: [scheduled, confirmed, arrived, in_progress, completed, cancelled, no_show]
        reason:
          type: string
        notes:
          type: string
    AppointmentListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── QUEUE SCHEMAS ────────────────────────────────────────────────────────
...