# lib/api-zod/src/generated/types/appointmentListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hospital Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { Appointment } from './appointment';

export interface AppointmentListResponse {
  data: Appointment[];
  total: number;
  page: number;
  limit: number;
}